refactor: extract vcs detection into src/api/vcs.ts
init.ts had grown a 200-line tail of VCS-specific functions (provider/
host map, Vercel + Netlify env detection, URL parsing, and the
applyVcsResourceAttributes shell). Move all of it into a dedicated
src/api/vcs.ts module. init.ts now references VCS in exactly two
places: the import and the one-line call from
initializeResourceAttributes.

Public API and behaviour are unchanged. Tests are untouched (they
exercise applyVcsResourceAttributes through init() as integration
coverage — that surface stays the same).
